backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
SQLAlchemy==2.0.41
python-dotenv==1.0.0
Authlib==1.3.0
requests==2.31.0
google-auth==2.26.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.2.0
